+++ packages/legend-query-builder/src/stores/milestoning/QueryBuilderProcessingTemporalMilestoningImplementation.ts
@@ -42,6 +42,8 @@ export class QueryBuilderProcessingTemporalMilestoningImplementation extends Que
         ),
       );
     }
+    // Show the parameter panel because we populate paramaters state with milestoning parameters
+    this.milestoningState.queryBuilderState.setShowParametersPanel(true);
   }
 
   processGetAllParamaters(parameterValues: ValueSpecification[]): void {